ai-build/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
dataclasses-json==0.6.7
docker==7.1.0
faiss-cpu==1.9.0
fastapi==0.115.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.59
langchain-text-splitters==0.3.8
langsmith==0.1.147
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
networkx==3.4.2
numpy==1.26.4
orjson==3.10.18
packaging==24.2
pillow==11.2.1
propcache==0.3.1
pydantic==2.9.2
pydantic-settings==2.9.1
pydantic_core==2.23.4
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==3.1.1
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.38.6
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.67.1
transformers==4.44.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.30.6
yarl==1.20.0
